+++ workspaces/jfrog-artifactory/plugins/jfrog-artifactory/src/translations/it.ts
@@ -26,7 +26,8 @@ const jfrogArtifactoryTranslationIt = createTranslationMessages({
   messages: {
     'page.title': 'Repository Artifactory di JFrog: {{image}}',
     'table.searchPlaceholder': 'Filtro',
-    'table.labelRowsSelect': 'Righe',
+    'table.pagination.showResults': 'Mostra {{count}} risultati',
+    'table.pagination.rangeLabel': '{{start}} - {{end}} di {{total}}',
     'table.columns.version': 'Versione',
     'table.columns.repositories': 'Repository',
     'table.columns.manifest': 'Manifesto',