+++ workspaces/kiali/plugins/kiali-common/src/types/ServiceList.ts
@@ -27,7 +27,7 @@ export interface ServiceOverview {
   name: string;
   cluster?: string;
   istioSidecar: boolean;
-  istioAmbient: boolean;
+  isAmbient: boolean;
   additionalDetailSample?: AdditionalItem;
   labels: { [key: string]: string };
   ports: { [key: string]: number };