+++ workspaces/airbrake/plugins/airbrake-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-airbrake-backend
 
+## 0.23.0
+
+### Minor Changes
+
+- 81095bf: Backstage version bump to v1.52.0
+
 ## 0.22.0
 
 ### Minor Changes