+++ workspaces/cicd-statistics/plugins/cicd-statistics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-cicd-statistics
 
+## 0.18.0
+
+### Minor Changes
+
+- bbed3c8: Backstage version bump to v1.52.0
+
 ## 0.17.0
 
 ### Minor Changes