+++ workspaces/github/plugins/search-backend-module-github-discussions/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-search-backend-module-github-discussions
 
+## 0.14.0
+
+### Minor Changes
+
+- 7e98f23: Backstage version bump to v1.52.0
+
+### Patch Changes
+
+- Updated dependencies [7e98f23]
+  - @backstage-community/plugin-github-discussions-common@0.14.0
+
 ## 0.13.0
 
 ### Minor Changes